Updated to use size_t instead of unsigned

+++ C++/stack.hh
@@ -2,6 +2,7 @@
 #ifndef STACK_HH
 #define STACK_HH
 
+#include <cstddef>
 #include <cstring>
 #include <exception>
 #include <stdexcept>
@@ -10,7 +11,7 @@ template <typename T>
 class Stack
 {
 public:
-	Stack(unsigned size = 0)
+	Stack(size_t size = 0)
 	{
 		this->_base     = nullptr;
 		this->_pointer  = nullptr;
@@ -53,10 +54,10 @@ public:
 	 * If the new size is smaller than the already allocated size, the data is copied
 	 * from the bottom of the stack up to the new size
 	 */
-	void allocate(unsigned size)
+	void allocate(size_t size)
 	{
-		T *memory       = new T[size];
-		unsigned offset = this->size();
+		T *memory     = new T[size];
+		size_t offset = this->size();
 
 		if(this->_base != nullptr)
 		{
@@ -79,12 +80,12 @@ public:
 		this->_pointer = this->_base;
 	}
 
-	unsigned size(void) const
+	size_t size(void) const
 	{
-		return (unsigned)(this->_pointer - this->_base);
+		return (size_t)(this->_pointer - this->_base);
 	}
 	
-	unsigned capacity(void) const
+	size_t capacity(void) const
 	{
 		return this->_capacity;
 	}
@@ -103,17 +104,17 @@ public:
 		return *(--this->_pointer);
 	}
 	
-	void pick(unsigned n)
+	void pick(size_t n)
 	{
 		if(this->size() <= n)
 			throw std::underflow_error("Stack underflow");
 		this->push(this->_pointer[-(int)(n+1)]);
 	} 
 
-	void roll(unsigned n)
+	void roll(size_t n)
 	{
 		this->pick(n);
-		for(int i = -(int)(n+2); i <= -2; ++i)
+		for(long long i = -(long long)(n+2); i <= -2; ++i)
 		{
 			this->_pointer[i] = this->_pointer[i+1];
 		}
@@ -180,7 +181,7 @@ public:
 private:
 	T *_base;
 	T *_pointer;
-	unsigned _capacity;
+	size_t _capacity;
 };
 
 #endif /* STACK_HH */
